requirements.txt

fastapi==0.109.0
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.4.2
requests==2.31.0